+++ modules/special_functions/builtin/cpp/betaincBuiltin.cpp
@@ -23,27 +23,49 @@
 // License along with this program. If not, see <http://www.gnu.org/licenses/>.
 // LICENCE_BLOCK_END
 //=============================================================================
-#include "NelsonGateway.hpp"
-#include "gammaBuiltin.hpp"
 #include "betaincBuiltin.hpp"
+#include "Error.hpp"
+#include "BetaIncomplete.hpp"
+#include "OverloadFunction.hpp"
+#include "ClassName.hpp"
 //=============================================================================
 using namespace Nelson;
 //=============================================================================
-const std::wstring gatewayName = L"elementary_functions";
-//=============================================================================
-static const nlsGateway gateway[] = {
-    { "gamma", (void*)Nelson::SpecialFunctionsGateway::gammaBuiltin, 1, 1,
-        CPP_BUILTIN_WITH_EVALUATOR },
-    { "betainc", (void*)Nelson::SpecialFunctionsGateway::betaincBuiltin, 1, 3,
-        CPP_BUILTIN_WITH_EVALUATOR },
-
-};
-//=============================================================================
-NLSGATEWAYFUNC(gateway)
-//=============================================================================
-NLSGATEWAYINFO(gateway)
-//=============================================================================
-NLSGATEWAYREMOVE(gateway)
-//=============================================================================
-NLSGATEWAYNAME()
+ArrayOfVector
+Nelson::SpecialFunctionsGateway::betaincBuiltin(
+    Evaluator* eval, int nLhs, const ArrayOfVector& argIn)
+{
+    ArrayOfVector retval;
+    if (argIn.size() < 2 || argIn.size() > 4) {
+        Error(ERROR_WRONG_NUMBERS_INPUT_ARGS);
+    }
+    if (nLhs > 1) {
+        Error(ERROR_WRONG_NUMBERS_OUTPUT_ARGS);
+    }
+    bool bSuccess = false;
+    if (eval->mustOverloadBasicTypes()) {
+        retval = OverloadFunction(eval, nLhs, argIn, "betainc", bSuccess);
+    }
+    if (!bSuccess) {
+        bool needToOverload = false;
+        bool isLower = true;
+        if (argIn.size() == 4) {
+            std::wstring tail = argIn[3].getContentAsWideString();
+            if (tail.compare(L"upper") == 0 || tail.compare(L"lower") == 0) {
+                if (tail.compare(L"upper") == 0) {
+                    isLower = false;
+                } else {
+                    isLower = true;
+                }
+            } else {
+                Error(_("Wrong value of the fourth argument 'upper' or 'lower' expected."));
+            }
+        }
+        retval.push_back(BetaIncomplete(argIn[0], argIn[1], argIn[2], isLower, needToOverload));
+        if (needToOverload) {
+            retval = OverloadFunction(eval, nLhs, argIn, "betainc");
+        }
+    }
+    return retval;
+}
 //=============================================================================