+++ modules/types/src/cpp/Dimensions.cpp
@@ -93,7 +93,8 @@ Dimensions::getMax()
     return maxL;
 }
 //=============================================================================
-indexType& Dimensions::operator[](indexType i)
+indexType&
+Dimensions::operator[](indexType i)
 {
     if (i >= maxDims) {
         Error(_("Too many dimensions! Current limit is") + " " + std::to_string(Nelson::maxDims)
@@ -110,14 +111,16 @@ indexType& Dimensions::operator[](indexType i)
 }
 //=============================================================================
 indexType
-Dimensions::getAt(indexType i)
+Dimensions::getAt(indexType i, bool checkLength)
 {
     if (i >= maxDims) {
         Error(_("Too many dimensions! Current limit is") + " " + std::to_string(Nelson::maxDims)
             + ".");
     }
-    if (i >= length) {
-        Error(_("Invalid dimension position."));
+    if (checkLength) {
+        if (i >= length) {
+            Error(_("Invalid dimension position."));
+        }
     }
     return data[i];
 }